move childrens for WrapperHashtagHistory
