+++ src/App.js
@@ -1,47 +1,24 @@
 import React, { useState } from "react";
 import "./App.css";
 
-import InputCheckbox from "./objects/InputCheckbox";
 import AboutPage from "./objects/AboutPage";
+import WrapperHashtagHistory from "./objects/WrapperHashtagHistory";
 
 import HeaderGame from "./components/HeaderGame";
-import HashtagGame from "./components/HashtagGame";
 import HeaderInternal from "./components/HeaderInternal";
 import ProfileUser from "./components/ProfileUser";
-import HistoryGame from "./components/HistoryGame";
-import WrapperHashtagHistory from "./objects/WrapperHashtagHistory";
 
 const App = () => {
     const [activeAbout, setActiveAbout] = useState("");
-    const [history, setHistory] = useState(["Start"]);
-    const [active, setActive] = useState(false);
 
     const handleClickAdd = () => setActiveAbout("-active");
-
     const handleClickRemove = () => setActiveAbout("");
-
-    const addHistory = player => {
-        setHistory(old => [...old, `Adicionou ${player.toUpperCase()}`]);
-    };
-
-    const ShowHideHistory = () => setActive(old => !!!old);
-
-    const changeHistory = key => setHistory(old => old.slice(0, key + 1));
 
     return (
         <main id="main" className="app">
             <HeaderGame onClick={handleClickAdd} />
 
-            <WrapperHashtagHistory active={active}>
-                <HashtagGame callback={addHistory} />
-                <InputCheckbox
-                    onClick={ShowHideHistory}
-                    id="show"
-                    value="show"
-                    content="Mostrar eventos"
-                />
-                <HistoryGame history={history} onClick={changeHistory} />
-            </WrapperHashtagHistory>
+            <WrapperHashtagHistory />
 
             <AboutPage className={activeAbout}>
                 <HeaderInternal onClick={handleClickRemove} />
