+++ src/objects/LogGame/styles.css
@@ -1,12 +1,13 @@
 .log-game {
-  background-color: var(--color-tertiary);
-  border-radius: 2em;
-  padding: var(--space-smallest) var(--space-middest);
-  font-size: var(--size-min);
-  color: var(--color-primary);
+    background-color: var(--color-tertiary);
+    border-radius: 2em;
+    padding: var(--space-smallest) var(--space-middest);
+    font-size: var(--size-min);
+    color: var(--color-primary);
+    cursor: pointer;
 }
 
 .log-game.-end {
-  background-color: var(--color-draw);
-  text-transform: uppercase;
+    background-color: var(--color-draw);
+    text-transform: uppercase;
 }
